refacto(backend) : refacto du module admin
