+++ src/network/projet_quizz/backend/src/admin/admin.module.ts
@@ -1,11 +1,25 @@
 import { Module } from '@nestjs/common';
 import { PrismaModule } from '../prisma/prisma.module';
 import { AdminController } from './admin.controller';
-import { AdminService } from './admin.service';
+import {
+  AdminService,
+  FormatV1Strategy,
+  JsonExporterService,
+  MergeOrchestratorService,
+  SqlExporterService,
+  SqlReplaceService,
+} from './services';
 
 @Module({
   imports: [PrismaModule],
   controllers: [AdminController],
-  providers: [AdminService],
+  providers: [
+    SqlExporterService,
+    JsonExporterService,
+    SqlReplaceService,
+    FormatV1Strategy,
+    MergeOrchestratorService,
+    AdminService,
+  ],
 })
 export class AdminModule {}
