+++ src/network/projet_quizz/backend/src/admin/dto/admin.dto.ts
@@ -1,0 +1,10 @@
+import { IsString, MinLength } from 'class-validator';
+
+export class ImportDatabaseSqlDto {
+  @IsString()
+  @MinLength(1)
+  script!: string;
+
+  @IsString()
+  confirm!: string;
+}
